Module Hmap.Make

Functor for heterogeneous maps whose keys hold information of type Key_info.t

Parameters

Key_info : KEY_INFO

Signature

Keys

type 'a key

The type for keys whose lookup value is of type 'a.

module Key : sig ... end with type 'a info = 'a Key_info.t

Keys.

module Make : functor (Value_info : VALUE_INFO) -> sig ... end