Module Mimic__Hmap

Heterogeneous value maps.

3.3.1-55-g300e98add - homepage

Functorial interface

The functorial interface allows to associate more information to the keys. For example a key name or a key value pretty-printer.

module type KEY_INFO = sig ... end

The type for key information.

module type VALUE_INFO = sig ... end

type ('a, 'b) teq =

| Teq : ('a, 'a) teq

module type S = sig ... end

Output signature of the functor Make

module Make : functor (Key_info : KEY_INFO) -> S with type 'a Key.info = 'a Key_info.t

Functor for heterogeneous maps whose keys hold information of type Key_info.t