Module Mimic

module Mirage_protocol : sig ... end

type flow = private ..

The type for flows. A flow represents the state of a single reliable stream stream that is connected to an endpoint.

include Mirage_flow.S with type flow := flow and type error = [ `Msg of string | `Not_found | `Cycle ]

type error = [

| `Cycle

| `Msg of string

| `Not_found

]

val pp_error : error Fmt.t

type nonrec write_error = private [>

| Mirage_flow.write_error

]

val pp_write_error : write_error Fmt.t

val read : flow -> ( Cstruct.t Mirage_flow.or_eof, error ) Stdlib.result Lwt.t

val write : flow -> Cstruct.t -> ( unit, write_error ) Stdlib.result Lwt.t

val writev : flow -> Cstruct.t list -> ( unit, write_error ) Stdlib.result Lwt.t

val close : flow -> unit Lwt.t

type ctx

The type for contexts. It's a heterogeneous map of values to help mimic to instantiate a new flow via resolve.

type 'edn value

The type for witnesses whose lookup value is of type 'edn.

module Fun : sig ... end

val make : name:string -> 'edn value

make ~name is a new witness.

val add : 'edn value -> 'edn -> ctx -> ctx

add w v ctx is ctx with w bound to v.

val get : 'edn value -> ctx -> 'edn option

get w ctx is the value of w's binding in ctx, if any.

val fold : 'edn value -> ( 'k, 'edn option Lwt.t ) Fun.args -> k:'k -> ctx -> ctx

val merge : ctx -> ctx -> ctx

val empty : ctx

empty is the empty context.

type ('edn, 'flow) protocol

val register : ?priority:int -> name:string -> (module Mirage_protocol.S with type endpoint = 'edn and type flow = 'flow) -> 'edn value * ( 'edn, 'flow ) protocol

register ?priority ~name (module Protocol) registers the given Protocol into the internal global Mimic's state as a possible transmission protocol available via resolve.

?priority is used to help mimic to choose between multiple solutions according to the given context. Mimic will choose the lower-priority solution.

name helps the end-user to know which solution mimic will dynamically via log outputs.

register returns 2 values:

a witness as the required value to initiate a transmission via the given Protocol implementation

a protocol which can help the end-user to destruct a flow to its structural type via repr.

module type REPR = sig ... end

val repr : ( 'edn, 'flow ) protocol -> (module REPR with type t = 'flow)

repr protocol gives a module definition with an OCaml constructor to help the end-user to destruct the structural type of a given flow:

module Protocol
  : Mirage_protocol.S with type flow = Lwt_unix.file_descr

let edn, protocol = Mimic.register ~name:"protocol" (module Protocol)
module R = (val (Mimic.repr protocol))

let () = Mimic.resolve ~ctx >>= function
  | Ok (R.T lwt_unix_file_descr) -> ...
  | ...

val resolve : ctx -> ( flow, [> error ] ) Stdlib.result Lwt.t

resolve ctx tries to instantiate a flow from the given ctx.

type edn =

| Edn : 'edn value * 'edn -> edn	(*
The type of a value and its witness.
*)

type (_, _) refl =

| Refl : ( 'a, 'a ) refl

val equal : 'a value -> 'b value -> ( 'a, 'b ) refl option

equal a b returns a proof that a and b are structurally equal.

val unfold : ctx -> ( edn list, [> `Cycle ] ) Stdlib.result Lwt.t

unfold ctx applies any functions available into the given ctx and and possible to compute according to available values and return a list of what these functions return.

It's useful to do an introspection of what mimic does when it resolves the given ctx. From that and equal, the user is able to introspect what mimic generated and which protocol it is able to instantiate then.

al:resolve} is:
    {[
      let resolve ctx =
        unfold ctx >>= function
        | Ok lst -> connect lst
        | Error _ as err -> Lwt.return err
    ]}

val connect : edn list -> ( flow, [> error ] ) Stdlib.result Lwt.t

connect values tries to instantiate a flow from given values and registered protocols (see register).

module Merge (A : sig ... end) (B : sig ... end) : sig ... end